This hefty submission adds a new connection type to Maestro: Local Connection
The Local Connection is an implementation of IServerConnection that wraps the desktop implementation of the MapGuide API (http://mg-desktop.googlecode.com). For purposes of convenience, the mg-desktop binaries (and FDO) and base CS-Map dictionary files are included in this submission. These were built from the recent trunk revisions of MapGuide, FDO and CS-Map. The FDO binaries included here is a subset of the actual FDO binaries, since I lack the ability to build certain providers.

Also included is a new addin that registers this connection type on startup and a LocalConfigure console application that writes out a suitable Platform.ini that's needed to initialize the desktop MapGuide API

git-svn-id: http://svn.osgeo.org/mapguide/trunk/Tools/Maestro@6087 a442ad82-b40c-0410-ae44-e6d7bd6e4a36