#1708: Have ODBC child editors expose a RequestDocumentReset event. Only the DSN child editor raises this event. Thus changing the DSN will invalidate and reset the configuration document. The DSN child editor UI has been changed to allow for prompting and notification of reset configuration documents.
Also it turns out we also need to internalise the ctors of RuntimeMapGroup as well because our planned future alternate implementation of RuntimeMapLayer that wraps MgLayer needs to be associated with MgLayerGroup. So we need to be able to have derivatives of RuntimeMapGroup, thus creation also needs to be a provider specific responsibility. Thus new CreateMapGroup() methods have been added to the IMappingService interface (one overload for each previously public ctor in RuntimeMapGroup), with default implementations in PlatformConnectionBase

git-svn-id: http://svn.osgeo.org/mapguide/trunk/Tools/Maestro@6312 a442ad82-b40c-0410-ae44-e6d7bd6e4a36