#2571: Fix Maestro ProviderTemplate tool for MGOS 3.0 support.
Because the Maestro API now is built with VS2015/Roslyn, it breaks our ProviderTemplate tool as we can no longer shell out to csc.exe to compile the required Maestro provider source files.

We have to do this programmatically with the new Roslyn compiler.

Since we're switching to Roslyn, this submission also removes various options from the tool which are either obsolete or no longer worth supporting:

 - Path to csc.exe (no longer required)
 - Assembly layout (we're assuming the split assembly layout introduced with 2.2 onwards)

While this pretty much should fix the tool, compiling against MGOS 3.0 assemblies exposed issues in the provider source that have to be fixed (the usage of various deprecated APIs that have been removed in 3.0)

Another unresolved issue is that I've yet to figure out how strong naming works in Roslyn, which we'll have to figure out because we're referencing strong-named MapGuide assemblies.

git-svn-id: http://svn.osgeo.org/mapguide/trunk/Tools/Maestro@8761 a442ad82-b40c-0410-ae44-e6d7bd6e4a36