Multi-target MaestroAPI and ObjectModels projects (net461 and netstandard2.0).
This was needed as a workaround to Maestro (full framework) wanting the full framework net45 version of GeoAPI instead of the netstandard one, which caused a ripple effect wrt strong naming (net45 GeoAPI is strong named, netstandard is not). Although we referenced a netstandard GeoAPI (not strong named), something in MSBuild decided that it wanted the net45 version instead (which is strong named). At runtime, this meant we get a FileLoadException when loading GeoAPI as we referenced the netstandard version, but MSBuild put the net45 strong named version into the output path instead. Since it appears MSBuild will give a full framework application a full framework class library over a netstandard one, we might as well multi-target the MaestroAPI and ObjectModels project so that the MSBuild computed dependency chain for Maestro.csproj is full framework all the way down. As a result, we've reinstated strong naming for MaestroAPI and ObjectModels for the net461 target. The netstandard2.0 target remains unsigned.