#51: Add XYZ tile seeding support. Unlike the MapGuide tile seeding mode, this has a simplified set of parameters:
 - A URL template with {x}, {y} and {z} placeholders
 - A required lat/long bounding box of the region to seed

In the process of implementing the XYZ tile seeding support, it was discovered that the initial tile seeding abstractions still had MapGuide-isms attached. For example, the XYZ tile seeding does not care about a resource ID or base layer group name, but these are required in the tile seeding interfaces that the XYZ tile walker and service classes have to implement. For now, we just return empty strings.