Fix basic line style trashed on cancellation of the edit style dialog. With basic line styles, it is effectively a collection of stroke objects. When retaining an original copy of the line style, we are keeping a reference to the original stroke list. This doesn't work (unlike other feature style edits) as any in place edits in the dialog will modify the stroke collection reference, trashing the original result.
The fix is to copy the original stroke items into a temporary list, and on the "rollback" action, to restore the original stroke items (on our temporary list) back to the original line style. In place edits can do whatever they want, the "rollback" action will restore the original strokes.

Fixes #42