Gracefully handle invalid resource state with open editors.
For example, opening a file-based Feature Source whose referenced data file does not exist. Previously Maestro would crash out (due to the underlying exception thrown in an application idle event), now we can at least re-throw a more meaningful exception and render the editor in question in a state that clearly advertises that it is not usable.