Fix a long standing annoyance where setting WMS bounds in Layer Definition metadata will set the coordinate system to EPSG:???? requiring you to manually enter in the EPSG code.
That is stupid, we're crunching spatial extents already, and in the process we get the coordinate system WKT. So we can just get the corresponding EPSG code from that via the Coordinate System Catalog meaning the user should not even have to manually enter in the required EPSG code. Maestro should automatically infer it most of the time!