Add new user preference for specifying the base URL of a mapguide-react-layout installation.
If set, this will unlock a whole set of mapguide-react-layout preview URLs available for the Web and Flexible Layout editors. The existing viewer URL text box has been changed to a combo box to allow for the desired viewer URL to be previewed. As this now prevents manually selecting the public URL for copy/paste, a new button is placed beside it to copy the given URL to the clipboard.

Fixes #20